+++ .gitignore
@@ -1,4 +1,19 @@
+target/
+*.class
+*.log
 .settings
-bin
 .classpath
 .project
+.target
+
+.metadata
+bin/
+tmp/
+*.tmp
+*.bak
+*.swp
+*~.nib
+local.properties
+.settings/
+.loadpath
+.recommenders
