+++ src/main/java/br/com/alura/edigi/dao/AutorDao.java
@@ -1,0 +1,9 @@
+package br.com.alura.edigi.dao;
+
+import br.com.alura.edigi.modelo.Autor;
+
+public interface AutorDao {
+
+	public void adiciona(Autor autor);
+
+}
