+++ src/main/java/br/com/alura/edigi/dao/CategoriaDao.java
@@ -1,0 +1,9 @@
+package br.com.alura.edigi.dao;
+
+import br.com.alura.edigi.modelo.Categoria;
+
+public interface CategoriaDao {
+
+	public boolean adiciona(Categoria categoria);
+
+}
