+++ src/main/resources/database.properties
@@ -1,0 +1,3 @@
+database.url=jdbc:mysql://localhost/edigi?useTimezone=true&serverTimezone=UTC
+database.user=root
+database.password=1q1a2w2s@@MYSQL
